cli: extract stdin helpers shared by new and append into stdin.go
